add eslint-config-airbnb peer dependencies as dependencies
Unfortunately the eslint module resolver cannot resolve the plugin
packages relative to the configuration file. It seems that it is trying
to use the packages from this preset instead of using the packages from the
project folder.

I think this might be related to neutrino-lint-base having eslint as a
dependency, instead of a peerDependency.